fix type errors and rename all backend files
