+++ server/src/types/group.types.ts
@@ -1,4 +1,4 @@
-import { Base, Timestamps } from './commonType';
+import { Base, Timestamps } from './common.types';
 
 export interface Group extends Base, Timestamps {
   name?: string;
