+++ server/src/utils/validators.ts
@@ -1,4 +1,4 @@
-export const validateMedia = (data: [string]) => data.length >= 5;
+export const validateMedia = (data: string[]) => data.length >= 5;
 
 export const validateTime = (data: string) => {
   const regex = /^(0[0-9]|1[0-2]):[0-5][0-9] (AM|PM)$/;
